requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
startapp==0.2.0.0
static3==0.7.0
urllib3==1.26.12
whitenoise==6.2.0
